+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Dependency changes
 
 * Require R 4.2 or higher (for native pipe and placeholder)
+* Add `withr` to Suggests for use with tests
 
 ## Bugfixes